+++ jpegenc/subblocks/huffman/__init__.py
@@ -1,6 +1,6 @@
 
 from __future__ import absolute_import
-import csv
+
 from .dc_rom import dc_rom
 from .ac_rom import ac_rom
 from .ac_cr_rom import ac_cr_rom
@@ -9,7 +9,10 @@ from .dc_cr_rom import dc_cr_rom
 from .huffman import (HuffmanCntrl, ImgSize, HuffmanDataStream,
                       BufferDataBus, huffman)
 
+from .tablebuilder import build_huffman_rom_tables
+
 __all__ = [
     'dc_rom', 'ac_rom', 'dc_cr_rom', 'ac_cr_rom', 'HuffmanCntrl',
-    'ImgSize', 'HuffmanDataStream', 'huffman', 'BufferDataBus'
+    'ImgSize', 'HuffmanDataStream', 'huffman', 'BufferDataBus',
+    'build_huffman_rom_tables'
 ]
